Added new database feature
Now the data will be saved to the database and we do not need to fetch
from the server each time. We will also save all search position from
where user was standing when doing the search. By doing this we know
that if user was standing to a previous location then all location exist
in the database already and we do not need to fetch from server.